Move `PerasWeightSnapshot` to separate module
It makes sense to use this without using an entire PerasCertDB, so decouple
these.

It might be nice to rename PerasCertDB-bench, but doesn't seem like a priority.
