+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Peras/Weight.hs
@@ -1,0 +1,45 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+
+module Ouroboros.Consensus.Peras.Weight
+  ( PerasWeightSnapshot (..)
+  , boostedWeightForPoint
+  , boostedWeightForFragment
+  ) where
+
+import Data.Map.Strict (Map)
+import qualified Data.Map.Strict as Map
+import GHC.Generics (Generic)
+import NoThunks.Class
+import Ouroboros.Consensus.Block
+import Ouroboros.Network.AnchoredFragment (AnchoredFragment)
+import qualified Ouroboros.Network.AnchoredFragment as AF
+
+newtype PerasWeightSnapshot blk = PerasWeightSnapshot
+  { getPerasWeightSnapshot :: Map (Point blk) PerasWeight
+  }
+  deriving stock (Show, Eq)
+  deriving Generic
+  deriving newtype NoThunks
+
+boostedWeightForPoint ::
+  forall blk.
+  StandardHash blk =>
+  PerasWeightSnapshot blk -> Point blk -> PerasWeight
+boostedWeightForPoint (PerasWeightSnapshot weightByPoint) pt =
+  Map.findWithDefault mempty pt weightByPoint
+
+boostedWeightForFragment ::
+  forall blk.
+  HasHeader blk =>
+  PerasWeightSnapshot blk ->
+  AnchoredFragment blk ->
+  PerasWeight
+boostedWeightForFragment weightSnap frag =
+  -- TODO think about whether this could be done in sublinear complexity
+  foldMap
+    (boostedWeightForPoint weightSnap)
+    (blockPoint <$> AF.toOldestFirst frag)
